Move assertions.js inside testring; add .eslintrc.json to that folder
