+++ packages/core/lib/testing/.eslintrc.json
@@ -1,0 +1,10 @@
+{
+  "overrides": [
+    {
+      "files": ["./Test.js"],
+      "env": {
+        "mocha": true
+      }
+    }
+  ]
+}
